// web/src/types/services/purchase-additional-cost-category/PurchaseAdditionalCostCategoryRequest.ts
export interface PurchaseAdditionalCostCategoryReadAnyPaginateRequest {
  with_trashed: boolean;
  company_id: string;
  search?: string | null;
  include_id?: string;
  refresh: boolean;
  page: number;
  per_page: number;
}

export interface PurchaseAdditionalCostCategoryReadAnyGetRequest {
  with_trashed: boolean;
  company_id: string;
  search?: string | null;
  include_id?: string;
  refresh: boolean;
  limit: number;
}